Move awaits to a separate file and fix await commitment
